Fix intercept handling
