+++ pyfixest/estimation/formula/transforms/fixed_effects_encoding.py
@@ -1,4 +1,9 @@
+import functools
+import itertools
+
 import pandas as pd
+from formulaic.parser import DefaultOperatorResolver
+from formulaic.parser.types import Operator, OrderedSet
 from formulaic.utils.stateful_transforms import stateful_transform
 
 
@@ -13,3 +18,28 @@ def fixed_effect_interactions(*args, _state=None, _metadata=None, _spec=None):
         return data["__id__"]
 
     return data.merge(_state, on=data.columns, how="left")["__id__"]
+
+
+class _FixedEffectsOperatorResolver(DefaultOperatorResolver):
+    def __init__(self):
+        super().__init__()
+
+    @property
+    def operators(self) -> list[Operator]:
+        operators = [
+            operator for operator in super().operators if operator.symbol != "^"
+        ]
+
+        operators.append(
+            Operator(
+                symbol="^",
+                arity=2,
+                precedence=500,
+                associativity="left",
+                to_terms=lambda *term_sets: OrderedSet(
+                    functools.reduce(lambda x, y: x * y, term)
+                    for term in itertools.product(*term_sets)
+                ),
+            )
+        )
+        return operators
