+++ pyfixest/estimation/formula/utils.py
@@ -4,6 +4,8 @@ from enum import Enum
 
 import pandas as pd
 
+from pyfixest.errors import FormulaSyntaxError
+
 
 def _str_split_by_sep(string: str, separator: str = "+") -> list[str]:
     """
@@ -78,3 +80,53 @@ class _MultipleEstimationType(Enum):
 _MULTIPLE_ESTIMATION_PATTERN = re.compile(
     rf"\b({'|'.join(me.name for me in _MultipleEstimationType)})\b\(.+\)"
 )
+
+
+def _preprocess(formula: str) -> str:
+    formula = _preprocess_fixest_instrumental_variable(formula)
+    formula = _preprocess_fixest_multiple_dependents(formula)
+    return formula
+
+
+def _preprocess_fixest_instrumental_variable(formula: str) -> str:
+    """Convert fixest-style instrumental variable syntax to formulaic.
+    Y ~ X1 | X2 ~ Z2 will be converted to Y ~ X1 + [X2 ~ Z2].
+    """
+    parts = re.split(r"\s*\|\s*", formula)
+    main = parts.pop(0)
+    instrumental_variables = [part for part in parts if "~" in part]
+    if len(instrumental_variables) > 1:
+        raise FormulaSyntaxError()
+    elif instrumental_variables:
+        parts = [part for part in parts if part not in instrumental_variables]
+        formula_old = formula
+        formula = f"{main} + {' + '.join(f'[{iv}]' for iv in instrumental_variables)}"
+        if parts:
+            formula = f"{formula} | {' | '.join(parts)}"
+        warnings.warn(
+            "The fixest-style syntax for instrumental variable regressions is deprecated and will throw an error in a future version."
+            f"Instead of `{formula_old}` use `{formula}`",
+            DeprecationWarning,
+            stacklevel=2,
+        )
+    return formula
+
+
+def _preprocess_fixest_multiple_dependents(formula: str) -> str:
+    """Convert multiple dependent variables to multiple estimation syntax.
+    Y + Y2 ~ X1 + X2 will be converted to sw(Y, Y2) ~ X1 + X2.
+    """
+    if "~" not in formula:
+        raise FormulaSyntaxError()
+    dependent, rest = re.split(r"\s*~\s*", formula, maxsplit=1)
+    if "+" in dependent:
+        # Multiple dependent variables
+        formula_old = formula
+        formula = f"{_MultipleEstimationType.sw.name}({', '.join(_str_split_by_sep(dependent, separator='+'))}) ~ {rest}"
+        warnings.warn(
+            "Specifiying multiple dependent variables with `+` is deprecated and will throw an error in a future version."
+            f"Instead of `{formula_old}` use `{formula}`",
+            DeprecationWarning,
+            stacklevel=2,
+        )
+    return formula
